+++ src/main/java/data_access/CreateCommentDataAccessInterface.java
@@ -3,7 +3,7 @@ package data_access;
 import entity.Comment;
 import org.bson.types.ObjectId;
 
-public interface CommentDataAccessInterface {
+public interface CreateCommentDataAccessInterface {
     void addComment(Comment comment);
 
     ObjectId getLoggedInUserId();
