+++ src/main/java/use_case/create_comment/application_business_rules/CreateCommentInputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.comment.application_business_rules;
+package use_case.create_comment.application_business_rules;
 
 public interface CreateCommentInputBoundary {
     void execute(CreateCommentInputData createCommentInputData);
