+++ src/main/java/use_case/create_comment/application_business_rules/CreateCommentInputData.java
@@ -1,4 +1,4 @@
-package use_case.comment.application_business_rules;
+package use_case.create_comment.application_business_rules;
 
 import org.bson.types.ObjectId;
