+++ src/main/java/use_case/create_comment/application_business_rules/CreateCommentInteractor.java
@@ -1,6 +1,6 @@
-package use_case.comment.application_business_rules;
+package use_case.create_comment.application_business_rules;
 
-import data_access.CommentDataAccessInterface;
+import data_access.CreateCommentDataAccessInterface;
 import entity.Comment;
 import entity.CommentFactory;
 import org.bson.types.ObjectId;
@@ -9,11 +9,11 @@ import java.util.Arrays;
 import java.util.List;
 
 public class CreateCommentInteractor implements CreateCommentInputBoundary {
-    final CommentDataAccessInterface commentDataAccessObject;
+    final CreateCommentDataAccessInterface commentDataAccessObject;
     final CreateCommentOutputBoundary commentPresenter;
     final CommentFactory commentFactory;
 
-    public CreateCommentInteractor(CreateCommentOutputBoundary commentPresenter, CommentDataAccessInterface commentDataAccessObject, CommentFactory commentFactory){
+    public CreateCommentInteractor(CreateCommentOutputBoundary commentPresenter, CreateCommentDataAccessInterface commentDataAccessObject, CommentFactory commentFactory){
         this.commentDataAccessObject = commentDataAccessObject;
         this.commentPresenter = commentPresenter;
         this.commentFactory = commentFactory;
