+++ src/main/java/view/CreateCommentView.java
@@ -1,7 +1,7 @@
 package view;
 
-import use_case.comment.interface_adapter.CreateCommentController;
-import use_case.comment.interface_adapter.CreateCommentState;
+import use_case.create_comment.interface_adapter.CreateCommentController;
+import use_case.create_comment.interface_adapter.CreateCommentState;
 
 
 import javax.swing.*;
